Move agent skill checks to dev tools
